Working version with type inference!
Fixes a couple of shallow bugs.
Adds copying of type variables.
Adds support for printing expressions with types, though expressions
in proofs do not have explicit types at this point, just type checking
as they are created.
Decorating expressions with type info would require a lot of copying of
Var (and other) objects that is not done now.