Add source step tracking for expressions introduced as assumptions. This means adding the property and modifying expression transformers to avoid copying except where required. Carry the property through in the "copy" method for convenience of renderers.
Make renderers and toString sensitive to a new trackSourceSteps
parameter.  This has been useful so far in debugging.

Detect when a step can be displayed with hypotheses, and present
it accordingly.

Make the Path parsing function support /main for convenience.

Fix unHyp to choose the right path.

Add support for hypotheses to useDefinition, apply, instForall.
Tidy up handling of paths in the "replace" rule.

Fix one hover-handler since the input to instForall can be top-level
or have hypotheses.

Add tests for new/changed functionality.